+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 
+## [0.14.6] - no yet released
+
+- Export ``isiterable`` and ``isasynciterable`` in ``mpservice.streamer``.
+- ``mpservice.threading.Thread.raise_exc`` was renamed to ``throw``.
+
+
 ## [0.14.5] - 2024-01-18
 
 - Customize ``Process`` and ``Thread`` classes in ``mpservice.mpserver`` to log unhandled BaseException, if any.